[finance_report_vision] Split extraction.py (1903-line god class) into mixins
Behaviour-preserving decomposition of the ExtractionService god class (25
methods, 1903 lines) into a package of responsibility mixins, each well under
1000 lines; public import surface unchanged.

- extraction/_base.py: shared constants, date parsing, ExtractionError.
- extraction/service.py: ExtractionService (core LLM parse path + class consts)
  composing the mixins.
- extraction/{_media,_coerce,_ocr,_brokerage,_csv}.py: responsibility mixins.
- extraction/__init__.py: re-exports + keeps `import httpx` so the existing
  httpx monkeypatch path resolves.

Test impact (test-only): tests that monkeypatched LLM primitives at the flat
module path (stream_ai_json/accumulate_stream/DeduplicationService/
validate_balance_per_currency) are repointed to the submodule where each is now
looked up (service for the JSON path, _csv for the CSV path). Four hard-coded
extraction.py path guards (secure_glm, observability) are made package-aware.
537 extraction tests pass; full collection clean.
