+++ apps/backend/tests/extraction/test_pdf_parsing.py
@@ -336,7 +336,7 @@ class TestParsingTimeout:
 
         from src.services.ai_streaming import AIStreamError
 
-        with patch("src.services.extraction.stream_ai_json") as mock_stream:
+        with patch("src.services.extraction.service.stream_ai_json") as mock_stream:
             mock_stream.side_effect = AIStreamError("Connection timed out")
 
             with pytest.raises(ExtractionError):
@@ -368,7 +368,7 @@ class TestGeminiRetry:
 
         from src.services.ai_streaming import AIStreamError
 
-        with patch("src.services.extraction.stream_ai_json") as mock_stream:
+        with patch("src.services.extraction.service.stream_ai_json") as mock_stream:
             mock_stream.side_effect = AIStreamError("HTTP 429: Rate limit exceeded")
 
             with pytest.raises(ExtractionError, match="429"):
